gh-pages auto publishes to gh-pages branch of users repo. Simple script to call gh-pages added for deploy, deploy-dev has replaced deploy for publishing to CFB page
