+++ scripts/deploy-personal.js
@@ -1,24 +1,6 @@
-// get remote repositories
-require('simple-git')()
-    .listRemote(['--get-url'], (err, data) => {
-        if (!err) {
-            console.log('Remote url for repository at ' + __dirname + ':');
-            console.log(data);
- 
-            // write remote data to package.json
-            var fs = require('fs');
-            var fileName = '../package.json';
-            var file = require(fileName);
+var ghpages = require('gh-pages');
+var path = require('path');
 
-            // data contains a newline char, removing before adding to json
-            file.homepage =  data.replace(/\n/g, '');
-
-            fs.writeFile(fileName, JSON.stringify(file, null, 2), function (err) {
-            if (err) return console.log(err);
-            console.log('writing to ' + fileName);
-            });
-        }
-    });
-
-
-
+// push the dev branch to the user's gh-pages branch
+// creates a gh-pages branch if one does not exist
+ghpages.publish('build');
